Added the -c|--clear-all-output option to ./replay.sh to clear all output files.
Before this commit, when calling ./sauron.py on a single file, it was
overwriting its output.txt file even if it was specified not to do so.
--> Now the problem is solved.

In this version of the project, every single fit is stored in a single
output.txt file and ./replay.sh add everything in the general
output.txt.
In this way one can handle every single peak separately.